+++ CHANGELOG.md
@@ -10,12 +10,17 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 - Displaying current user-agent on extension icon hover [#364]
 
+### Changed
+
+- The whitelist/blacklist modes toggle redesigned [#369]
+
 ### Fixed
 
 - Settings toggle for Black\White-List logic [#367]
 
 [#364]:https://github.com/tarampampam/random-user-agent/issues/364
 [#367]:https://github.com/tarampampam/random-user-agent/issues/367
+[#369]:https://github.com/tarampampam/random-user-agent/pull/369
 
 ## v3.14.2